1.5.82 release
* Imagine: Update C++ standard to C++26 and fix build issues
* Imagine: Add visit() member function for variant types using AddVisit CRTP class
* Imagine: Clean up decltype usage in various headers
* Imagine: Remove unused bundled FreeType library
* Imagine: Fix evaluating LD variable in makefiles too early
* NES.emu: Add/Update Lightful and Palightful palettes
* C64.emu, GBA.emu, GBC.emu, MD.emu, PCE.emu, Saturn.emu: Fix C++26 build errors due to deprecated arithmetic conversions on enums

+++ C64.emu/src/vice/resid/sid.cc
@@ -579,7 +579,7 @@ bool SID::set_sampling_parameters(double clock_freq, sampling_method method,
   if (method == SAMPLE_RESAMPLE || method == SAMPLE_RESAMPLE_FASTMEM)
   {
     // Check whether the sample ring buffer would overfill.
-    if (FIR_N*clock_freq/sample_freq >= RINGSIZE) {
+    if (double(FIR_N)*clock_freq/sample_freq >= double(RINGSIZE)) {
       return false;
     }
